download & error handling
