+++ requirements.txt
@@ -11,6 +11,7 @@ pandas==2.0.3
 pluggy==1.2.0
 pyarrow==13.0.0
 pytest==7.4.0
+pytest-mock=3.12.0
 python-dateutil==2.8.2
 pytz==2023.3
 reactivex==4.0.4
