+++ test_api.py
@@ -1,5 +1,5 @@
 import pytest
-from api import DessAPI
+from api import DessAPI, DessApiException
 import re
 from unittest.mock import patch
 
@@ -12,18 +12,36 @@ def test_instance(monkeypatch):
 def test_sign(test_instance):
     assert re.match(r"\w+", test_instance.sign())
 
-@pytest.fixture
-def mock_requests_get():
-    with patch('requests.get') as mock_get:
-        yield mock_get
-
-def test_function_to_test_with_params(mock_requests_get, test_instance):
+def test_call_api_with_success_with_right_params(test_instance):
     url = "http://web.dessmonitor.com/public/"
     other_params = {"param2": "value2"}
+    
 
-    test_instance.get(**other_params)
+    with patch('requests.get') as mock_get:
+        mock_get.return_value.status_code = 200
+        mock_get.return_value.headers = {"Content-disposition": "attachment; filename=test.csv"}
+        test_instance.get(**other_params)
 
-    # Check that requests.get was called with the right parameters
-    args, kwargs = mock_requests_get.call_args
+    args, kwargs = mock_get.call_args
     assert args[0] == url
     assert 'sign' in kwargs['params'] or 'token' in kwargs['params']
+
+def test_raise_exception_when_token_expired(test_instance):
+    other_params = {"param2": "value2"}
+
+    with patch('requests.get') as mock_get:
+        mock_get.return_value.status_code = 200
+        mock_get.return_value.headers = {"Content-type": "application/json"}
+        with pytest.raises(DessApiException) as excinfo:
+            test_instance.get(**other_params)
+    assert excinfo.match(r"^API returned no data")
+    
+
+def test_raise_exception_when_api_down(test_instance):
+    other_params = {"param2": "value2"}
+
+    with patch('requests.get') as mock_get:
+        mock_get.return_value.status_code = 500
+        with pytest.raises(DessApiException) as excinfo:
+            test_instance.get(**other_params)
+    assert excinfo.match(r"^API returned status code")
